Adicionando o id Cliente na Grid de pesquisa; Atualizar a grid de clientes ao carregar a tela;
Conflicts:

	src/br/com/artefino/ordermanager/client/ArteFinoOrderManagerConstants.java

+++ src/br/com/artefino/ordermanager/client/model/ClienteRecord.java
@@ -22,7 +22,7 @@ public class ClienteRecord extends ListGridRecord {
 	}
 
 	public void setId(Long attribute) {
-		setAttribute(ID, attribute);
+		setAttribute(ID, attribute.longValue());
 	}
 
 	public void setNome(String attribute) {
